update addNoise and add noise function in StereoBase

+++ include/noise.hpp
@@ -4,6 +4,6 @@
 
 namespace cp
 {
-	CP_EXPORT void addNoise(cv::InputArray src, cv::OutputArray dest, const double sigma, const double solt_papper_ratio = 0.0, const int seed=0);
+	CP_EXPORT void addNoise(cv::InputArray src, cv::OutputArray dest, const double sigma, const double solt_papper_ratio = 0.0, const uint64 seed = 0);
 	CP_EXPORT void addJPEGNoise(cv::InputArray src, cv::OutputArray dest, const int quality);
 }
